AMPS-539 Fix the build
git-svn-id: https://studio.atlassian.com/svn/AMPS/trunk@129940 2c54a935-e501-0410-bc05-97a93f6bca70

+++ maven-amps-plugin/src/main/java/com/atlassian/maven/plugins/amps/product/AmpsProductHandler.java
@@ -123,8 +123,11 @@ public abstract class AmpsProductHandler implements ProductHandler
 
             // Get rid of "studio-test-resources.zip", which is the homeZip that was used
             // when we started Amps.
-            String originalHomeZip = this.getTestResourcesArtifact().getArtifactId() + ".zip";
-            FileUtils.deleteQuietly(new File(snapshotDir, originalHomeZip));
+            if (this.getTestResourcesArtifact() != null)
+            {
+                String originalHomeZip = this.getTestResourcesArtifact().getArtifactId() + ".zip";
+                FileUtils.deleteQuietly(new File(snapshotDir, originalHomeZip));
+            }
 
 
             // Proceed to replacements
